renamed sub reducers to get rid of smurf naming, i.e. reducer- in the file name is redundant due to the parent directory. 2) renamed select-square to selection as reducers should be named after the part of the state they manage rather than any specific action, 3) named file and import consistently in lowercase (camelCase) which as a bonus allowed simplified object syntax
