+++ src/index.js
@@ -5,97 +5,7 @@ import { Provider } from 'react-redux'
 import App from './components/App'
 import checkersApp from './reducers'
 
-const store = createStore(checkersApp,
-  { board: [
-        [
-          {piece: null},
-          {piece: "black"},
-          {piece: null},
-          {piece: "black"},
-          {piece: null},
-          {piece: "black"},
-          {piece: null},
-          {piece: "black"}
-        ]
-      ,
-        [
-          {piece: "black"},
-          {piece: null},
-          {piece: "black"},
-          {piece: null},
-          {piece: "black"},
-          {piece: null},
-          {piece: "black"},
-          {piece: null}
-        ]
-      ,
-        [
-          {piece: null},
-          {piece: "black"},
-          {piece: null},
-          {piece: "black"},
-          {piece: null},
-          {piece: "black"},
-          {piece: null},
-          {piece: "black"}
-        ]
-      ,
-        [
-          {piece: null},
-          {piece: null},
-          {piece: null},
-          {piece: null},
-          {piece: null},
-          {piece: null},
-          {piece: null},
-          {piece: null}
-        ]
-      ,
-        [
-          {piece: null},
-          {piece: null},
-          {piece: null},
-          {piece: null},
-          {piece: null},
-          {piece: null},
-          {piece: null},
-          {piece: null}
-        ]
-      ,
-        [
-          {piece: "red"},
-          {piece: null},
-          {piece: "red"},
-          {piece: null},
-          {piece: "red"},
-          {piece: null},
-          {piece: "red"},
-          {piece: null}
-        ]
-      ,
-        [
-          {piece: null},
-          {piece: "red"},
-          {piece: null},
-          {piece: "red"},
-          {piece: null},
-          {piece: "red"},
-          {piece: null},
-          {piece: "red"}
-        ]
-      ,
-        [
-          {piece: "red"},
-          {piece: null},
-          {piece: "red"},
-          {piece: null},
-          {piece: "red"},
-          {piece: null},
-          {piece: "red"},
-          {piece: null}
-        ]
-    ]
-  });
+const store = createStore(checkersApp);
 
 render(
   <Provider store={store}>
