+++ src/reducers/board.js
@@ -1,0 +1,93 @@
+const defaultState = [
+      [
+        {piece: null},
+        {piece: "black"},
+        {piece: null},
+        {piece: "black"},
+        {piece: null},
+        {piece: "black"},
+        {piece: null},
+        {piece: "black"}
+      ]
+    ,
+      [
+        {piece: "black"},
+        {piece: null},
+        {piece: "black"},
+        {piece: null},
+        {piece: "black"},
+        {piece: null},
+        {piece: "black"},
+        {piece: null}
+      ]
+    ,
+      [
+        {piece: null},
+        {piece: "black"},
+        {piece: null},
+        {piece: "black"},
+        {piece: null},
+        {piece: "black"},
+        {piece: null},
+        {piece: "black"}
+      ]
+    ,
+      [
+        {piece: null},
+        {piece: null},
+        {piece: null},
+        {piece: null},
+        {piece: null},
+        {piece: null},
+        {piece: null},
+        {piece: null}
+      ]
+    ,
+      [
+        {piece: null},
+        {piece: null},
+        {piece: null},
+        {piece: null},
+        {piece: null},
+        {piece: null},
+        {piece: null},
+        {piece: null}
+      ]
+    ,
+      [
+        {piece: "red"},
+        {piece: null},
+        {piece: "red"},
+        {piece: null},
+        {piece: "red"},
+        {piece: null},
+        {piece: "red"},
+        {piece: null}
+      ]
+    ,
+      [
+        {piece: null},
+        {piece: "red"},
+        {piece: null},
+        {piece: "red"},
+        {piece: null},
+        {piece: "red"},
+        {piece: null},
+        {piece: "red"}
+      ]
+    ,
+      [
+        {piece: "red"},
+        {piece: null},
+        {piece: "red"},
+        {piece: null},
+        {piece: "red"},
+        {piece: null},
+        {piece: "red"},
+        {piece: null}
+      ]
+  ];
+
+export default function(state=defaultState, action){
+  return state;
+}
