--- src/reducers/reducer-board.js
@@ -1,3 +1,0 @@
-export default function(state=null, action){
-  return state;
-}
